Medical QA/requirements.txt
# Core Dependencies
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
# For GPU support, use: faiss-gpu>=1.7.4
pdfplumber>=0.9.0
transformers>=4.30.0
accelerate>=0.20.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Model Inference
llama-cpp-python>=0.2.0
# For GPU support: pip install llama-cpp-python --extra-index-url https://abetlen.github.io/llama-cpp-python/whl/cu118

# API Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# Fine-tuning
peft>=0.4.0
bitsandbytes>=0.41.0
datasets>=2.14.0

# NLP Utilities
nltk>=3.8.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
